+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Change Log
 
-## Unreleased
+## [0.0.15] - 2023-07-10
 
 - Fixed
   - Fixed false positive `DOC402` when `yield` statements are in a block within
-    a child function
+    a nested function
+  - Fixed false positives when `return` and `raise` statements are in a block
+    within a nested function
+- Full diff
+  - https://github.com/jsh9/pydoclint/compare/0.0.15...0.0.15
 
 ## [0.0.14] - 2023-07-05